Split index validation helpers
